requirements.txt
pyyaml==5.3.1
torch==1.10.0
torchvision==0.8.2
einops==0.3.2